likes/tests/requirements/19.txt
Django==1.9.6
pytz==2016.4
django-secretballot==0.6.0
